- Moved all item translation stuff to the plugin menu (This should fix those new NP1015 errors) - Clicking on the item translation plugin (translator.dll) will show the configuration box - Added some more code to plugin updates - Added support for plugin configurations (optional) - Fixed some bugs - Rewrote some redundant strings - Improved the initial plugin installation process - Fixed some item patch configuration bugs

+++ PSO2 Tweaker/frmItemConfig.vb
@@ -3,8 +3,13 @@ Imports PSO2_Tweaker.My
 
 Public Class FrmItemConfig
     Private Sub btnSave_Click(sender As Object, e As EventArgs) Handles btnSave.Click
+        If cmbToggleKey.Text = "Please choose a key" Then
+            MsgBox("Please choose a key to toggle the item patch!")
+            Exit Sub
+        End If
         Dim pso2Launchpath = Program.Pso2RootDir
-        File.Delete(pso2Launchpath & "\translation.cfg")
+        If File.Exists(pso2Launchpath & "\translation.cfg") Then File.Delete(pso2Launchpath & "\translation.cfg")
+
 
         Dim splitKey As String() = DirectCast(cmbToggleKey.SelectedItem, String).Split("("c)
         Dim data As String()
